Refactoring analyzer after split up
Rename `codechecker` module to `codecheckeranalyzer` under analyzer.
