+++ analyzer/cmd/parse.py
@@ -18,7 +18,8 @@ import sys
 
 from plist_to_html import PlistToHtml
 
-from codechecker import analyzer_context
+from codechecker_analyzer import analyzer_context
+
 from libcodechecker import suppress_handler
 from libcodechecker import logger
 from libcodechecker import util
@@ -26,7 +27,6 @@ from libcodechecker import plist_parser
 from libcodechecker.skiplist_handler import SkipListHandler
 from libcodechecker.output_formatters import twodim_to_str
 from libcodechecker.report import Report, get_report_path_hash
-# TODO: This is a cross-subpackage reference...
 
 LOG = logger.get_logger('system')
