+++ analyzer/codechecker_analyzer/__init__.py
