+++ analyzer/codechecker_analyzer/analysis_manager.py
