+++ analyzer/codechecker_analyzer/analyzer.py
