+++ analyzer/codechecker_analyzer/analyzer_context.py
