+++ analyzer/codechecker_analyzer/analyzers/__init__.py
