+++ analyzer/codechecker_analyzer/analyzers/analyzer_base.py
