+++ analyzer/codechecker_analyzer/analyzers/analyzer_clang_tidy.py
