+++ analyzer/codechecker_analyzer/analyzers/analyzer_clangsa.py
