+++ analyzer/codechecker_analyzer/analyzers/analyzer_types.py
