+++ analyzer/codechecker_analyzer/analyzers/config_handler_clang_tidy.py
