+++ analyzer/codechecker_analyzer/analyzers/config_handler_clangsa.py
