+++ analyzer/codechecker_analyzer/analyzers/ctu_triple_arch.py
