+++ analyzer/codechecker_analyzer/analyzers/result_handler_base.py
