+++ analyzer/codechecker_analyzer/analyzers/result_handler_clang_tidy.py
