+++ analyzer/codechecker_analyzer/buildlog/host_check.py
