+++ analyzer/codechecker_analyzer/ctu_manager.py
