+++ analyzer/codechecker_analyzer/gcc_toolchain.py
