+++ analyzer/codechecker_analyzer/pre_analysis_manager.py
