+++ analyzer/codechecker_analyzer/statistics_collector.py
