+++ analyzer/codechecker_analyzer/tidy_output_converter.py
